+++ src/domain/builder/builders/markdown.builder.ts
@@ -13,12 +13,8 @@ const parseMarkdown = async (markdown: string) => {
 };
 
 export const markdownBuilder: IBuilder = async (cache: ICache, file: IFile) => {
-  const nameParts = file.name.split(".");
-  nameParts[nameParts.length - 1] = "html";
-
   const newFile: IFile = {
     ...file,
-    name: nameParts.join("."),
     extension: "html",
     raw: await parseMarkdown(file.raw.toString())
   };