+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Changed
 
 - Updated node to 12.13 LTS.
+- File metadata name doesn't include extension.
 
 ## Fixed